requirements.txt
pytest==6.1.1
requests==2.24.0
responses==0.12.0
python-dotenv==0.14.0
python-telegram-bot==13.0
selenium==3.141.0